+++ CHANGES.rst
@@ -67,6 +67,8 @@ New Features
 
 - Table viewer "active row" selection to control data visible in other viewers. [#4279, #4322, #4350]
 
+- Allow markers plugin table to create other viewer types besides Scatter (i.e. Table or Histogram). [#4331]
+
 - Add ability to specify coordinate frame and equinox in catalog loader. [#4207]
 
 - Table viewer API & UI to add/rename/remove columns. [#4282, #4343]
@@ -98,7 +100,15 @@ New Features
 - Updated documentation for the application and caching, updated the jdaviz logo,
   and improved the .dmg installation process [#4337]
 
+- Coordinate frame and equinox are responsive to changes in ra / dec in catalog
+  loader. [#4347]
+
 - Improve UI appearance for URL loader. [#4354]
+
+- Adds search to extension menu in loaders when there are more than 3 extensions available. [#4372]
+
+- Generate a Gaussian curve from Line Analysis plugin results and optionally add to the 1D Spectrum viewer. [#4374]
+
 
 Mosviz
 ^^^^^^
@@ -122,6 +132,8 @@ Bug Fixes
 - Fixed closing a viewer not deleting the associated backend objects. [#4317]
 
 - Fix color picker not showing correctly. [#4332]
+
+- Fix error when typing while renaming data or subset. [#4359]
 
 Mosviz
 ^^^^^^
@@ -149,7 +161,9 @@ Other Changes and Additions
 Bug Fixes
 ---------
 
-- fix case where file drop loader would not show importer options. [#4303]
+- Fix case where file drop loader would not show importer options. [#4303]
+
+- fix URL loader not showing in UI. [#4361]
 
 - Fix image importer support for Roman L3 mosaic files. [#4309]
 
@@ -161,6 +175,8 @@ Bug Fixes
   caused an index error from glue. [#4333]
 
 - Avoid triggering a 2D spectrum-related warning for NIRISS images. [#4342]
+
+- Adjust data menu behavior to allow for showing more icons when the viewer is short. [#4352]
 
 Mosviz
 ^^^^^^